Reduce interrupts from VDMA engine
The Xilinx VDMA driver always enables interrupts which causes an IRQ to
trigger on each frame. Nothing is actually using that, but it cannot be
disabled in the driver. Set the "coalesc" field in the xilinx_vdma_config
struct to its maximum value to at least reduce the number of interrupts.